rename(chore): providers 위치 변경
