+++ src/app/components/common/Toast/Toast.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from "@storybook/react";
 import Toast from "./Toast";
-import { ToastType } from "@/app/types/ToastTypes";
+import { ToastType } from "@/types/ToastTypes";
 
 const meta: Meta<typeof Toast> = {
   title: "Components/Toast",
